hash.bash can now hash directories
Method for doing so is still a trial.  There are various ways for
hashing a directory, and they don't all come up with the same output.
See notes at line 112 for a more thorough explanation, as well as my
hopes in trying to cover multiple systems, and also make this easy to
implement in other languages.